Upload source files with our preferred hash
We are moving away from md5 for hashes.

This gets the upload command to just upload the files with the
configured lookaside hash.

It only requires a corresponding change to the lookaside cache when we
change the actual config to something else than 'md5'.

That will come in a future commit, though.