Simplify handling of our custom certificates
The property loader is an overly-complex solution to cache the results
of getting a property.

In the case of our certificates, the call is not very expensive so we
don't even need to cache it at all.

However, given that pyrpkg now has a cached_property decorator which
does just that, we might as well use it.

Signed-off-by: Mathieu Bridon <bochecha@daitauha.fr>