lookaside: Use our new download path
We are moving to a new hash type for source tarballs in Fedora.

As part of this process, we are moving to a new download URL where the
hash type is a part of the path to the file to download.

As of today, Wed Jul 15 2015, all the source files are available at that
new URL.

That means it is time to start downloading from that new URL now.

If upstream pyrpkg ever moves its default to the new URL format, we can
even drop this downstream implementation altogether.

Signed-off-by: Mathieu Bridon <bochecha@daitauha.fr>