Move to the new sources file format
This is a pre-requisite to moving away from MD5 sum, as fedpkg will need
to know which hashtype was used for a give line in the sources file.

Upstream pyrpkg has supported this for long enough now (it was part of
rpkg-1.32 which is in all Fedora and EPEL branches) that we can make the
move in fedpkg.

Signed-off-by: Mathieu Bridon <bochecha@daitauha.fr>