enable target dest for rawhide to have trailing pieces
We want to make changes in fedora to enable build gating, either
for signing, automated QA or both.  we attempted to implement
automated signing and hit a snag. When the target pointed to
f26-pending the disttag for teh build nvr ended up being .fc26-pending
in order to enable flexibility we should split on - and take the value
from the -

Signed-off-by: Dennis Gilmore <dennis@ausil.us>