Remove deprecated call to platform.dist
As documented in Python documentation, platform.dist is deprecated
since version 2.6 and platform.linux_distribution should be used in new
code. Meanwhile, Python 2.6 is already the minimum version fedpkg works
with. It makes sense to remove the call to platform.dist at current
stage.

Tests are added and new code for writing and running tests are also
added.

This patch also fixes a bug that is original mydist is referenced
before assignment.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>