Automatically request module for non-standard branches.
This is half of fedrepo_req#129.

Whenever a user requests a non-standard branch, we will also submit a request
for a matching module with matching stream and matching SLs.

In a subsequent changeset, we want to make the admin tool automatically
populate new module repos with a useful modulemd starter template.

Signed-off-by: Ralph Bean <rbean@redhat.com>