Use base_module in clone_config
When cloning a module with a namespace, the `module` variable contains
the namespace prefix.  As used in `clone_config` (specifically, as
currently used in `bz.default-component` and `sendemail.to`), the
namespace prefix is not intended to be part of the value.

For example:

    $ fedpkg clone rpms/git
    Cloning into 'git'...
    ...
    $ cd git && git config --get-regexp 'bz.default-component|sendemail.to'
    bz.default-component rpms/git
    sendemail.to rpms/git-owner@fedoraproject.org

Use base_module for these values which strips the namespace prefix from
module.

Resolves: #230
Signed-off-by: Todd Zullinger <tmz@pobox.com>