Handle Bodhi login automatically
This refers to what Bodhi CLI errorhandler does. The reason to handle
login automatically is, in fedpkg implementation, list_overrides is
called pior to save_overrides and send_request. That causes Bodhi
cookies is stored, and then it results in AuthError during subsequent
call, e.g. to save_overrides.

So, the result behavior is similar with bodhi CLI, that is password is
required to be entered only once until it is unavailable.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>