Add option --namespace to command request-branch
This patch adds option --namespace to command request-branch, which
could be used with --repo together. Previously, packagers could request
a branch with global option --module-name in either form name or
namespace/name. --module-name has been replaced with options --name and
--namespace which should also be usable for request-branch as well.

As a result, packagers could request a branch with either global
options:

  fedpkg --name name [--namespace namespace] request-branch branch

or command level options

  fedpkg request-branch --repo name [--namespace namespace] branch

The first form is for backwards compatibility in a way and the second
form is the recommended way to use in practice which would be more
straightforward.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>