Greenwave conf and support for gating validation
rpkg will validate the gating.yaml file if there is one in the repo. It
will ask to Greenwave for that, when a build command is issued.
The command will fail if the gating.yaml file is misconfigured. The user
will have the chance to disable the check adding a parameter.
With this change fedpkg will have the default configuration to query the
Greenwave API.

Signed-off-by: Giulia Naponiello <gnaponie@redhat.com>