Resolve Jenkins unittests failing
Disabling Jenkins unittests for Python2.6, because it is not supported
by the Jenkins system anymore. To continue running unittests I also
had to change the approach to install dependent libraries - not to use
sitepackages. Libraries that are needed are installed by pip directly.
It leads to skipping of some tests that require bodhi-client.

JIRA: COMPOSE-3826
Signed-off-by: Ondrej Nosek <onosek@redhat.com>