Line up descriptions for better code readability
Multi-line descriptions of fedpkg commands were lined up to
the beginning of the line and it was not so readable. textwrap.dedent
was used to line descriptions up like normal string literals.
Width of the description column was originally 70 characters, with
this change you need to add a fixed number of spaces
on the beginning.

Signed-off-by: Ondrej Nosek <onosek@redhat.com>