Clone config customization for namespaces
After the 'clone' operation, a git config is updated with 'git-bz'
config values. Now, these values can be customized based on the
namespace of the cloned repository. Each namespace has its own
key 'clone_config_<namespace>' in config section. For namespaces,
that missing this config ('tests' namespace is a good example),
any of these values aren't updated.
A repository has to be namespaced ('distgit_namespaced = True')
for the config to be applied.
Functional change is part of the rpkg, and needs to be released
separately.

JIRA: COMPOSE-2638
Related: #231

Signed-off-by: Ondrej Nosek <onosek@redhat.com>